refactor: Table 패키지 분리
