+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -16,12 +16,12 @@ import kitchenpos.menu.repository.MenuRepository;
 import kitchenpos.order.application.OrderService;
 import kitchenpos.order.repository.OrderLineItemRepository;
 import kitchenpos.order.repository.OrderRepository;
-import kitchenpos.dao.OrderTableRepository;
+import kitchenpos.table.repository.OrderTableRepository;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 import kitchenpos.domain.Price;
 import kitchenpos.domain.TableGroup;
 import kitchenpos.order.dto.request.OrderLineItemCreateRequest;
