+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -11,10 +11,10 @@ import static org.mockito.Mockito.times;
 import java.util.Arrays;
 import java.util.List;
 import kitchenpos.order.repository.OrderRepository;
-import kitchenpos.dao.OrderTableRepository;
+import kitchenpos.table.repository.OrderTableRepository;
 import kitchenpos.dao.TableGroupRepository;
 import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
 import kitchenpos.dto.request.TableGroupCreateRequest;
 import org.junit.jupiter.api.DisplayName;
